readyset-mysql: Make the writer in the on_init callback optional
Making the writer in this callback optional paves the way for correctly
handling the database field of the MySQL client handshake response
packet. If the client specifies a database in the handshake response
packet, we need to call on_init to set the upstream database and the
schema search path, but we don't want to explicitly send back a separate
response the way we would if the client issued a USE statement. By
making the writer optional, the caller can specify None and skip the
code that sends a response back to the client.

Change-Id: Iae1c652996751bc7b0d6747c2f6654bf1b7f5d83
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4122
Reviewed-by: Peter Farr <peter@readyset.io>
Tested-by: Buildkite CI