mysql-srv: Implement handling for database field of handshake response
Prior to this change, we would ignore it if/when the client specified a
database to connect to during the initial handshake. Switching databases
in other ways is already implemented in MySQL, so adding support for
this ended up being pretty straightforward.

Release-Note-Core: ReadySet now correctly handles MySQL clients setting
  the database in the initial connection handshake. (Previously this was
  ignored and the client would instead be connected to the default
  database specified in the `--upstream-db-url` CLI argument to
  ReadySet.)
Fixes: ENG-1294
Change-Id: I323e330cd6c427adf1f1500d82b3bc88098d8e4b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4123
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>