mysql: Fixed binary types being displayed as strings.
Today, we are storing MySQL binary type values with different `DfValue`
(i.e, values can be `DfValue::ByteArray`, or `DfValue::TinyText`).
However, when we wrote the results of a query, we weren't paying
attention to the actual `DfType`, and we were solely relying on
the  `Display` impl of `DfValue` for text values.

One possible fix for this could've been to just have the right `Display`
 impl for `DfValue::ByteArray` and coerce the value to byte array if the
 column we are writing into is of type binary.
The problem, however, is that the display format for these types (and
I'm going to guess that others will have similar problems as well)
change depending on the database or even configuration (Postgres has two
 ways of displaying byte arrays), so we should stop relying on the
 `Display` impl of `DfValue`.

So, for now, this commit introduces a `BinaryFormat` simple struct that
will print the bytes it stores in the same format as MySQL.
This can be easily reverted/removed when (and if) we decide to change
the way we are handling the display of values.

Refs: ENG-1776
Change-Id: Id2f1aa127c7cdad6153c4e023a6906413539688b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4130
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>