Pass along parse failure to correct error
When a user tries to add a query that can't be cached because it can't
be parsed, we currently return an incorrect error that the query by the
provided query id cannot be found. This is incorrect - we did find the
query by the supplied query id, but the query was unparseable.

This commit updates our behavior so that we send back an
UnparseableQuery error, which provides correct information to a user
trying to cache a query we couldn't parse.

Change-Id: I5ca8f7a9074a487b1836b0fc056da128819b8eb1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4144
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>