data: allow coercing any Unknown to another Unknown
There are scenarios where we might want to coerce a value with the
inferred type Unknown to an inferred result type of Unknown, for example
certain postgresql string-to-string functions where the output type is
dependent on the input type. Previously, `coerce_to` would fail if both
the input and the output type were Unknown - this commit changes that to
just make the coercion a no-op instead.

Change-Id: I5b6771917aaa88d20bed80cb7faeb75d887a57e4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4153
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>