postgres replicator: Stop replicating tables containing unsupported NUMERICs
Drop and ignore tables containing NUMERIC values with scales > 28. The
scale is the number of digits right of the decimal. Our current
implementation of NUMERIC is backed by rust_decimal, which only supports
scales up to 28, and furthermore has a bug whereby it will deserialize
larger scales from the postgres binary protocol, but panic when
serializing again to bincode.

Refs: ENG-2264
Change-Id: I74b23d8c8a060014ac768997bf69b13f3d750494
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4163
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>