Apply changes on replication-tables config on restart
Forces a resnapshot when ReadySet starts up. Since the
replication-tables parameter is handled in snapshotting, this applies
updates to the parameter when the server restarts.

Also creates a ReplicationAction::LogPosition event for the postgres
replicator if we are attempting to catch up and our current position
matches or exceeds the catchup position. The mysql replicator already
follows this logic.

The LogPosition message is necessary because the resnapshot process
advances the binlog, but tables that have already been snapshotted
report an earlier binlog offset. Without this change, the catch up
process would not complete if no new writes/DDL were reported.

This CL mostly fixes ENG-2282. However, there is another bug where if we
snapshot a table, remove it from replication-tables, and then add it
back to replication-tables, we will not resnapshot the table because we
already have a replication offset for that table.

Also removes an unused start_with_authority() function.

Change-Id: Ia1f259a4b4dbea1b30c644d2032af674539b1fe6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4168
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>