nom-sql: Make CREATE CACHE parsing fallible
Using the same technique of making the field in the AST a Result and
using the `parse_fallible` helper which we've used to make the parsers
for `CREATE VIEW` and `CREATE TABLE` fallible, make the `CREATE CACHE`
parser fallible as well. In this case, we're using it for better error
messages - previously if a user attempted to run a `CREATE CACHE`
statement with a query we couldn't parse, we'd send the full query
upstream (as we do for all queries that fail to parse) and then the user
would see an error message from the upstream database that it couldn't
parse the CREATE CACHE statement. Now, instead, we just return a
`ReadySetError::UnparseableQuery` with the query that failed to parse,
which results in a much better error message

Release-Note-Core: Improved error messages for CREATE CACHE statements
  with queries that ReadySet is unable to parse
Fixes: ENG-1794
Change-Id: I3c43015d925a6ce7955d9d4b92af670a0665778a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4174
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>