server: Allow configuring worker request timeout
Add a new command-line flag and env var for configuring the timeout for
requests made to workers. We've been seeing this time out in production,
and even though it's currently set to 30 minutes which is *quite* long,
we might want to be able to increase it anyway in case the timeouts in
prod are unavoidable.

Change-Id: I55b8a57f4aa73b470bf9933b207c80d4539e8a1a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4181
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>