treewide: Replace all uses of maplit with Hash{Map,Set}::from
Now that min const generics is stabilized, we have the ability to
construct HashMaps from an array of tuples and HashSets from an array of
values. Given that, we no longer need the `maplit` crate to get easy
single-expression construction of those two types. That means we can get
rid of maplit everywhere we're using it without losing much, and get the
benefit of removing a dependency and remaining closer to "regular" rust
overall.

Change-Id: I4592abe2dba73f1379eb49c9fabe536d20ce6045
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4184
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>