readyset: Default creds to the upstream db url
Default both the --username and --password flags for the readyset binary
to the corresponding value parsed out of the upstream db url if not
specified. For cases when they're identical, which is in practice the
majority of cases, this allows users to not specify the flags, which is
a nice boost to UX.

*technically* we're parsing the upstream URL twice here - once for the
credentials and once for the actual connection upstream - but since it's
all only happening once (at startup) I figured it wasn't worth the
additional complexity required to optimize it away.

Release-Note-Core: The --username and --password fields will now default
  to the username and password, respectively, in the upstream database
  URL if not set.
Fixes: ENG-2308
Change-Id: If621ec9e91fa7b19f9840e4318041fa133bb46e1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4191
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>