tracing: Switch to OTLP protocol for sending traces
This replaces the jaeger-specific backend with one that speaks OTLP.
OTLP is OpenTelemetry's standardized protocol for transferring
traces, which has quite wide support (jaeger, zipkin, aws, and many
more).

With these changes tracing is fully functional in testing, and
significantly more reliable at sending trace data than
opentelemetry_jaeger.

From benchmarking using read_benchmark_news_app_small.yaml,
comparing against the current upstream code (as % of upstream,
lower is better):
 * 0% sample rate:   100.88%
 * 1% sample rate:   102.08%
 * 5% sample rate:   102.09%
 * 10% sample rate:  103.72%
 * 25% sample rate:  105.71%
 * 50% sample rate:  102.94%
 * 75% sample rate:  110.22%
 * 100% sample rate: 120.03%

Having such low overhead for 0% sample rate opens the door to
running in prod at 0%, and then at runtime adjust the sample rate
(this would require additional work, but probably not that much).
This could be leveraged to implement things such as dynamically
adjusting the system's sampling rate (e.g. raising it when
investigating behavior occurring on an already-running instance).
Another option would be allowing different parts of the code base to
have different sampling rates, such as 100% for Snapshotting while
the rest of the system is running at 0%).

A few changes were made to the deny.toml file. A few crates were
added to the skip list for the duplicate detection (updating them
elsewhere in ReadySet looked like a non-trivial change), and
RUSTSEC-2022-0055 was added to the ignore list as we're not using
tonic in a mode (as a server) that would be impacted.

Change-Id: I53c0dadbfc77ca44946cdc7d7a190813eb6ffbdb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4194
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>