replicators: Add new vertical DDL test suite
This adds a new proptest suite for testing Postgres replication of DDL
events. We generate sequences of DDL operations and writes/deletes of
rows, run them against both ReadySet and Postgres, and confirm after
each step that the state of the tables match up between ReadySet and
Postgres.

Failing cases are shrunk by removing one element at a time (where
possible to do so without violating any preconditions that are necessary
for running later steps).

Near future work includes the addition of more test ops, such as
creating and dropping views, as well as adding support for more types of
ALTER TABLE statements.

Further future work may include optimization of the shrinking algorithm,
as discussed in the comments. Additionally, this test suite has already
found one bug (described in ENG-2548) which we are currently working
around by tweaking our use of the `eventually!` macro to retry after
panics from the Postgres client code. If/when that bug is fixed, we can
remove the workaround.

Change-Id: I39c34d02dc1c9a9839ad920619a24afd624082e0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4332
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>