controller: Add debug logging for the HTTP server.
We are currently dealing with an issue with Holaplex, where all RPC
calls are failing due to timeouts, which is causing ReadySet to restart.
We didn't find anything in the logs (no panics or errors), and upon
manually getting into the instance, we also couldn't get a reply from
the server when trying to query endpoints like `/simple_graph`. This
worked for other customer instances though.

Our current guess is that either the HTTP server is panicking/erroring
without logging (doubtful, as a panic there should kill the worker
thread, and we have logs for that scenario) or the HTTP server is
getting stuck forever handling an HTTP request, most likely a write
(since they are the only ones that can block the entire server).

This commit then only adds debug logs to our `external_request` method,
so that we can see which HTTP requests we are getting, and if we
successfully handled them or if they are blocked.

Refs: ENG-2352
Change-Id: I1d219e6fc72cb9a5b3e388e69d8da77636e617bb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4203
Reviewed-by: Nick Marino <nick@readyset.io>
Tested-by: Buildkite CI