readyset-server: lower received domain log level
This log will eventually be useful as an info log because we may need to
fighure out which server a domain is running on, but currently we are
just using standalone, so it makes sense to lower this to debug until we
have multiple servers. Lowering it makes the logs less noisy and reduces
logging costs, as these are particularly noisy.

Change-Id: I1fdc19791a9bdbf6a811582828f263e78da95841
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4208
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>