replicators: Ignore tables that err durring repl
In `NoriaAdapter::main_loop`, we already handle table-specific errors that
arise when handling replication actions in such a way that we abandon
replicating the table without giving up on replication entirely.
However, we don't correctly handle errors that arise when _parsing_
replication events.

This CL:
- Separates table-specific `WalError`s into a separate enum variant
- Converts those `WalError`s into `ReadySetError::TableError`s
- Appropriately handles `ReadySetError::TableError`s in
  `NoriaAdapter::main_loop` so we don't give up on replication entirely
  when a table-specific error occurs
- Adds failpoint testing

I only added support for Postgres in this CL.

Release-Note-Core: Fixed an issue that caused a Postgres WAL event that
  failed to parse from being applied. Tables that encounter such errors will
  transition to not being replicated rather than periodically retrying.
Fixes: ENG-2345
Change-Id: Iaaf78e2b72f11bdaceb2ab239490fb7b543e4cc9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4209
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Luke Osborne <luke.o@readyset.io>