Remove error!() in get_view()
We currently log an error in get_view() if we fail to lookup a view in
the controller. However, we call get_view() in cases where we want to
check for a view, but don't necessarily expect a view to exist. In these
cases, the error logging is superfluous and noisy.

This CL removes the error. We already emit a warning if the view is not
found, and will also return an error if we hit this error while running
without an upstream.

Change-Id: I8a9c6d465aecbfe135806889c499868263d6410a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4212
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI