adapter: Avoid panic in lookup in QueryStatusCache
Previously, to allow tracking the query status of both queries that
successfully parsed and queries that failed to parse, we stored a
`Query` enum in the key of the `DashMap` inside of `QueryStatusCache` -
then, to allow methods on QueryStatusCache to be overloaded to take
either failed or successful parses, we implemented `Borrow` for both
String and ViewCreateRequest for Query, panicking if the variant was
incorrect. This panic was always incorrect when used to overload hash
map lookups, however - if both a failed parse and a sucessful parse
happened to hash to the same bucket, lookups would fall back to
borrowing and comparing for direct equality, which would then hit the
panic in the borrow impl.

I iterated quite a bit to try to hit the correct workaround for this,
and eventually landed on the best path forward being to have *two*
separate maps inside of `QueryStatusCache`, one that stores successful
parses and one that stores failed parses. Then, we can get the same kind
of overloading *outside* of `QueryStatusCache` by using a new
`QueryStatusKey` trait, which due to weirdness in the way DashMap's lock
guard types work needs its methods to take *closures* over optional
references (mutable or otherwise) to the values.

Unfortunately, since the only way to trigger this is to have two keys
that hash to the same bucket, and hashing in Rust is randomly seeded,
this is impractical to test, but definitionally since the line that was
panicking in the original bug report is gone this will definitely fix
that bug.

Release-Note-Core: Fixed a bug where in certain cases, executing a query
  that ReadySet is unable to parse would cause a panic to be logged from
  a background thread.
Fixes: ENG-2349
Change-Id: I3fbf6029db40dcc60945086faa4cdae3e9fdd1c0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4215
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Tested-by: Buildkite CI