replicators: Grant permissions for readyset schema to public
Some hosted database platforms, most specifically supabase in this case,
need access to all schemas in the database, to eg replicate them to a
secondary. Previously, we'd just create the `readyset` schema, which is
where we install all the ddl triggers and functions etc., as the
connected user, but not grant any permissions on it, which apparently
"deeply angers" supabase. This just grants usage permissions to PUBLIC
for that schema as part of the ddl replication setup sql script, which
has been confirmed in the wild to make supabase happy.

Fixes: ENG-2356
Change-Id: I09d3b5a7282bade275d0fd7fe1404d8c2ba7cb8a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4221
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>