Support multiple ReadySet per upstream
Previously we could only support one ReadySet instance per upstream when
that upstream was postgres. This is because we hard coded the
replication slot name, which caused a collision if more than one
ReadySet instance was pointed at the same postgres database.

This commit updates our behavior to use the deployment name as the
replication slot name. This gives us a handle to allow for multiple
readyset instances to connect to the same upstream.

Change-Id: Id82aead19131cde0e603a44d4b460cec8a8d7c80
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4241
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>