replicators: Set REPLICA IDENTITY FULL on snapshotted tables
As part of the snapshot process, automatically set `REPLICA IDENTITY
FULL` on tables we're snapshotting that both don't have a primary key,
and don't already have a non-default replica identity set. Without this,
postgresql prevents any destructive writes (deletes or updates) to these
tables, just because ReadySet has started replicating from them, which
is a bad user experience. The downside here is that these tables use
more space in the replication log, but given that the user experience of
being blocked from deleting or updating a table is quite bad we've
decided that the tradeoff here is worth it overall.

Refs: ENG-2085
Change-Id: Ifef076e9c8474ec9d99b3d346d4e667a0a0be5cf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4248
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>