replicators: set replica identity for new tables
As part of the function to create ddl events for newly snapshotted
tables, check to see if they need their REPLICA IDENTITY set to
FULL (because they don't have a primary key, and don't already have
their replica identity configured), and if so, run the ALTER TABLE
command to make that happen.

We need to be a little extra careful here to prevent our ALTER TABLE
event trigger from running on this statement - since we use
`current_query` to get the ALTER TABLE statement that was actually run,
but that actually returns the outermost `CREATE TABLE` statement that
triggered this ALTER TABLE, meaning we'll *always* restart the
replicator task on every CREATE TABLE. To prevent this from happening,
we set a `LOCAL` variable which we check in the ALTER TABLE event
trigger, then unset it after the ALTER TABLE statement is done.

Release-Note-Core: ReadySet now automatically runs ALTER TABLE <table>
  REPLICA IDENTITY FULL on PostgreSQL tables that ReadySet is configured
  to replicate from, and which don't have primary keys or
  already-configured non-default replica identities. This avoids an
  issue where ReadySet replicating from these tables would cause
  PostgreSQL to return an error message for updates and deletes to these
  tables
Fixes: ENG-2085
Change-Id: I0bac207c6483a35240941ac49cc84bc0f71036e5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4249
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>