replicators: Don't fail if parsing alter table fails
The approach we're currently using for generating DDL events for ALTER
TABLE is to just call the `current_query()` function and put the result
on the DDL replication stream. As documented, this function returns
the[0]:

> text of the currently executing query, as submitted by the
> client (might contain more than one statement)

This can work *sometimes*, specifically if the client ran the ALTER
TABLE as the only statement within the submitted query, and if the ALTER
TABLE itself wasn't triggered by another query (eg as a result of an
event trigger or something like that), but if the actual current query
is not immediately parseable as an ALTER TABLE statement this will just
completely blow up, causing all sorts of other downstream issues.
Fortunately, we don't actually *have* to always parse the ALTER TABLE
statement (at least right now), since literally all the time we just
need the name of the table that was altered so we can drop it and
resnapshot it. We can get the name of the table just by querying
`pg_class`, without having to care about the current query, and fall
back on that if parsing the alter table fails.

This approach likely *won't* scale well once we start actually directly
handling certain classes of ALTER TABLE without requiring a resnapshot,
since it'll mean that those statements might require a resnapshot
incorrectly, but should hold us over for the time being until we find a
better way of replicating the ALTER TABLE statement.

[0]: https://www.postgresql.org/docs/current/functions-info.html

Change-Id: I8ca3c035f5f163e1acf4ad36c43a8ba448ea1fb4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4255
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>