replicators: Bring back test db re-creation
In 7993f9987 (replicators: Drop and recreate the test db before tests,
2022-09-07), we started dropping and re-creating the test db in between
tests. For some reason, in 37dddad23 (replicators: replicate all tables
in all schemas by default, 2022-09-14), we changed that to only drop and
re-create the *schema* named after the test db between tests. This meant
that for example the `readyset` schema, where we store the tables and
triggers we use to replicate DDL, wasn't getting properly cleaned up in
between tests, which I *think* might have been causing some issues with
tests failing in CI but not locally that I've been seeing.

Change-Id: I4ad59fd75aa0b2f5f9c85da3547168f2894c43a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4278
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>