replicators: Group by all non-aggregated cols in type triggers
These queries work fine locally, but it seems like there must be a
subtle difference in the way the pg_type view is implemented on some
versions of Postgres, because a customer is getting an error about the
typarray column needing to be included in the GROUP BY clause or used in
an aggregate function. Normally this rule can be ignored if the GROUP BY
columns include the primary key of the underlying table, since that
guarantees that each of the table's other columns will have a unique
value for each group, and I believe this is why the query works on the
versions of Postgres that we've been testing with.

I'm guessing that some versions of Postgres may not consider oid to be a
primary key for the pg_type view, so to guard against this (and to
hopefully fix the error the customer is seeing) this commit adds the
non-grouped columns for these queries to the GROUP BY list. At worst,
even if this somehow does not fix the issue for the customer in
question, this change should have no effect and the queries should
continue to return the same results, so this seems like a fairly safe
thing to try.

Release-Note-Core: Fixed an issue where the Postgres triggers that
  ReadySet installs for DDL replication could cause errors when
  attempting to create or alter types.
Fixes: ENG-2492
Change-Id: Idff7e34f6b12b7675d894c29b69c370d0ae88992
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4279
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>