Don't resnapshot on setting REPLICA IDENTITY
We're going to be automatically setting the REPLICA IDENTITY on existing
and new tables in the replicator - resnapshotting every time we do that
is quite problematic (it doesn't *quite* lead to an infinite loop, but
it does slow things down and I've seen it cause the replicator to stop
replicating changes entirely in test scenarios). Since we *know* setting
replica identity doesn't actually do anything to the underlying data, we
can get away with just making it not require a resnapshot at
all (effectively, we're just ignoring it here, which is safe for us to
do).

Refs: ENG-2085
Change-Id: Ida839842f428c24767e411e7e61f34d3c236f628
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4281
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>