replicators: fail gracefully when conn task panics
We spawn a tokio task on which our connection to the upstream Postgres
instance runs. Most of the time, this task returns gracefully with
either nothing or with an error. Occasionally, however, the task can be
cancelled or the task may panic, which produces a `JoinError`.
Previously, we were assuming the task would never produce a `JoinError`
and calling `unwrap()` to ignore it. This change introduces error
handling that returns a `ReadySetError::ReplicationFailed` error when
presented with a `JoinError`.

Release-Note-Core: Fixed an issue where the ReadySet replicator wasn't
  gracefully handling every kind of Postgres connection failure. Now, the
  replicator will always retry in the event of a lost connection to the
  upstream database.
Fixes: ENG-2489
Change-Id: Ife665636c564f1d079d9f5171beaae1c51716497
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4287
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>