Remove deployment name from replication slot
Currently, by including the deployment name in the replication slot, we
require a number of constraints on the deployment name that we don't
want to impose on users. By only using the server id in the replication
slot, we can predictably always grab a replication slot, and still
support multiple ReadySet instances connecting to the same upstream.

Change-Id: I374d5166834d00b0d8aa5a993e438b17ad9a4b97
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4306
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>