readyset-psql: Decode enums as indices, then convert
To be consistent with the approach introduced to fix enum arrays for
ENG-2525, instead of decoding enum values as labels directly like we
were doing in 78ad3d4d2 (psql-srv: Don't decode enums as indices,
2022-11-18), decode them as indices, then convert them to labels before
querying upstream. This is not only more consistent with what we *have*
to do for arrays, but it's also more efficient (since variant indices
use less memory than labels), and more consistent with how we represent
these values everywhere else.

Refs: ENG-2525
Refs: ENG-2122
Change-Id: I0258a06efc9b8d48b01805a8805da3faa8019715
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4311
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>