dataflow-state: Fix db path for deployments with a dot in their name
Previously we were using `Path::set_extension` to make the RocksDB
database files we open for persistent state end with `.db`. If the
filename doesn't *already* have a dot, this just has the effect of
adding a `.db` to the end of the path - but if the filename *does* have
a dot, eg because the deployment name contains a dot, it'll *replace*
everything after that dot with `.db`. Since the format we use for base
table filename prefixes is `<deployment>-<schema>-<name>-<shard>`, this
meant that for deployments with a dot in their name, *all* base table
database files would just end up being the part of the deployment name
before the dot followed by `.db` - then, for any deployment with more
than one table, rocksdb would fail at startup due to trying to open the
same db twice.

Instead of using `Path::set_extension`, this just adds the `.db` to the
end of the filename using `String::push_str`, which is what we wanted
anyway, but doesn't have this edge condition.

Fixes: ENG-2494
Change-Id: I5ed8dff51a692129b3b3a6428a055a8f5e0931a9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4316
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>