domain: Don't fail when requesting replays from dropped readers
Sometimes, if enough things are happening fast enough, a reader node can
be dropped while there are still RequestReaderReplay packets in-flight
to the domain containing that reader - in that case, we don't want
to *fail* the domain, we just want to log a warning and drop the
packet (upstream, we'll timeout and retry the query against the
upstream).

Generally, we probably want to revisit error handling within domains so
that certain classes of errors encountered handling packets don't break
the domain, but that's a larger area of work than just this specific
issue.

Fixes: ENG-2346
Change-Id: I3597c15814afab989f9fe8a1da9e92329918ef7a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4319
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>