readyset-psql: Add ignored test to reproduce client crash
The vertical DDL test found this sequence of operations causes a client
crash, which appears to be caused by ReadySet sending back invalid
response packets. In the messages that ReadySet sends back to the
client for the final SELECT, the RowDescription shows two columns (which
includes the newly added row from the ALTER TABLE) but the DataRow only
includes data for one column.

This commit adds a failing unit test to document this, and is marked
"ignore" for now, but can be un-ignored once the bug is fixed.

Refs: ENG-2548
Change-Id: I8f243b5c8fc6e340bc3d3ba14910cfe86443d23f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4320
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI