readyset-util: Allow patterns in eventually! closure-like syntax
The run_test/then_assert form of the `eventually` macro uses a
closure-like syntax where the result of the run_test block gets passed
in as an argument to the then_assert block. However, when I wrote this
code I didn't consider the use case where we might want to do a pattern
match on that argument. This change allows for usages like this, which
would've previously triggered a compilation error:

```
eventually!(run_test: {
    (23, 42)
},
then_assert: |(a, b)| {
    assert_neq!(a, b);
});
```

Change-Id: Iaa1afb7748b9d98e081074035a8a14c3a6835724
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4323
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI