ddl_vertical: Add (but disable) ALTER TABLE DROP COLUMN operation
Initially, adding this code introduced a tricky crash in the generator
code where we could drop a column and then try to generate a
"RenameColumn" op on a table with no columns. Fixed this by tweaking the
RenameColumn generator to only be used if there are tables with columns.

Note that we're disabling this for now because this operation triggers
known failures due to the issues described in ENG-2548, where the
response packets don't match up and we get a client error. However,
unlike with AddColumn, we can't easily work around this via
`eventually!` because the corresponding error with DropColumn is
detected by the client code, and the connection is dropped.

We could establish a new connection for each postcondition check, but
that would slow down the test significantly, and doesn't seem worth it
right now. Once ENG-2548 is fixed, we can see about adding this
operation back in.

Change-Id: Ib1d64dd6d40ed61c8696fb83ba607352f89d2f43
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4391
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>