readyset-data: Test backwards compatibility of deserialization
Add a test suite, similar to the existing
`config_backwards_compatibility` test, for testing the backwards
compatibility of deserialization of `DfValue`. In this case, this is
also coupled to a `SERDE_VERSION` constant, which we will serialize to
persisted state and compare against on startup, to determine if we've
made any backwards incompatible changes to deserialization since that
state was written (in which case we'd throw it away and resnapshot).

Doing things this way, with a constant, avoids having to check that we
can deserialize *all* rows in persistent state, which would be required
if we went with a failure-only approach here (since it might be the case
that only some DfValue variants have backwards incompatibility).

Fixes: ENG-2564
Change-Id: I2b428cc7919c0815799bfb706517e960ae116179
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4336
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>