dataflow-state: Save and check DfValue serde version
Save the serde version used to serialize the values in persistent state
to the PersistentMeta, and check that against the SERDE_VERSION constant
on startup, returning an error if the serialized version is less than
our version. Currently this will just fail the domain, but this paves
the road to us being able to catch this error and use it to trigger a
resnapshot.

Refs: ENG-2564
Change-Id: I0feab48bc6857bda26c3645b65d1a770e8b59de4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4337
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>