dataflow-state: Use a custom error type for PersistentState
Some of the errors that can happen when creating a new persistent state
are things we want to handle directly - eg missing column families, or a
mismatched serde version. Some of the errors, on the other hand, are
truly unrecoverable - such as IO errors, etc (eg if the current user
doesn't have permission to write to the db directory, for example).
Rather than just putting everything in ReadySetError::Internal, this
commit introduces a new `Error` enum for persistent state, which we can
match on to handle those different kinds of errors differently.

Refs: ENG-2559
Change-Id: Ibfd47dbcec7dd0e18a5bdb51d430ac96fb170564
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4338
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI