dataflow-state: Drop and recreate state if creation errors
If opening a PersistentState returns a "temporary" error, handle the
error by deleting the database file and opening a new state (which will
then create an empty db, which will then resnapshot since it won't have
a ReplicationOffset). This essentially gets us zero-intervention
"backwards compatibility" for persisted state, at the cost of the
resnapshot if we do make a backwards-incompatible change.

Release-Note-Core: ReadySet will now automatically re-snapshot the
  snapshotted copy of base table state if necessary when upgrading to a
  new version.
Fixes: ENG-2560
Change-Id: I0e101de0267f833c4092865ec81b476fa79eef13
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4342
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>