treewide: shut down background tasks in tests
Now that we have the ability to gracefully await the shutdown of all of
the background tasks we spawn, we should be doing this at the end of
each test that spins up a controller. This change updates the
`TestBuilder` API such that a `ShutdownSender` is returned each time a
controller handle is returned. This requires users of `TestBuilder` to
either use the `ShutdownSender` or explicitly ignore it. By calling
`ShutdownSender::shutdown` at the end of each test, we ensure that the
test doesn't complete until all of the background tasks have been stopped.
This should fix a race condition we've been running into in our test
suite where, on occasion, abandoned background tasks panic because they
are still interacting with resources that are torn down by the test
suite.

Fixes: ENG-2314
Change-Id: I0eab9a9ba0626f9d2a428f9c38dd534ffc234e2c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4322
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>